Change design for Coin Performance
