+++ app/src/main/res/layout/coin_market_details.xml
@@ -8,14 +8,23 @@
     android:visibility="gone"
     tools:visibility="visible">
 
+    <io.horizontalsystems.bankwallet.modules.coin.CoinPerformanceView
+        android:id="@+id/coinPerformanceView"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="12dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
+
     <LinearLayout
         android:id="@+id/marketDataLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:layout_marginTop="12dp"
+        android:orientation="vertical"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/coinPerformanceView" />
 
     <io.horizontalsystems.views.SettingsView
         android:id="@+id/coinMarketsButton"
@@ -25,41 +34,6 @@
         app:layout_constraintTop_toBottomOf="@id/marketDataLayout"
         app:position="single"
         app:title="@string/CoinPage_CoinMarket" />
-
-    <View
-        android:id="@+id/border6"
-        android:layout_width="0dp"
-        android:layout_height="0.5dp"
-        android:layout_marginTop="12dp"
-        android:background="@color/steel_20"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/coinMarketsButton" />
-
-    <TextView
-        android:id="@+id/performanceTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="12dp"
-        android:layout_marginEnd="16dp"
-        android:text="@string/CoinPage_Performance"
-        android:textColor="@color/oz"
-        android:textSize="16sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/border6" />
-
-    <include
-        android:id="@+id/marketPerformance"
-        layout="@layout/coin_performance"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="24dp"
-        android:layout_marginEnd="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/performanceTitle" />
 
     <androidx.constraintlayout.widget.Group
         android:id="@+id/aboutGroup"
@@ -75,7 +49,7 @@
         android:background="@color/steel_20"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/marketPerformance" />
+        app:layout_constraintTop_toBottomOf="@id/coinMarketsButton" />
 
     <TextView
         android:id="@+id/aboutTitle"
@@ -122,8 +96,8 @@
         android:id="@+id/categoriesLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:layout_marginTop="20dp"
+        android:orientation="vertical"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/aboutTextToggle" />
 
@@ -131,8 +105,8 @@
         android:id="@+id/linksLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:layout_marginTop="12dp"
+        android:orientation="vertical"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/categoriesLayout" />
