--- app/src/main/res/layout/coin_performance.xml
@@ -1,205 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="24dp"
-    android:background="@drawable/background_round_16"
-    android:backgroundTint="@color/lawrence"
-    android:padding="16dp">
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/middleGuideline"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintGuide_percent="0.5013193" />
-
-    <TextView
-        android:id="@+id/changes1W"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="Changes (1W)"
-        android:textColor="@color/grey"
-        android:textSize="14sp"
-        app:layout_constraintEnd_toEndOf="@id/middleGuideline"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <View
-        android:id="@+id/changes1WBorder"
-        android:layout_width="0dp"
-        android:layout_height="0.5dp"
-        android:layout_marginTop="4dp"
-        android:layout_marginEnd="20dp"
-        android:background="@color/steel_20"
-        app:layout_constraintEnd_toEndOf="@id/middleGuideline"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/changes1W" />
-
-    <TextView
-        android:id="@+id/usd1W"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="USD"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/changes1WBorder" />
-
-    <TextView
-        android:id="@+id/usd1WPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="4dp"
-        android:text="+56,62%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/usd1W"
-        app:layout_constraintStart_toEndOf="@+id/usd1W"
-        app:layout_constraintTop_toTopOf="@+id/usd1W" />
-
-    <TextView
-        android:id="@+id/eth1W"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="ETH"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/usd1W" />
-
-    <TextView
-        android:id="@+id/eth1WPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="4dp"
-        android:text="+9,24%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/eth1W"
-        app:layout_constraintStart_toEndOf="@+id/eth1W"
-        app:layout_constraintTop_toTopOf="@+id/eth1W" />
-
-    <TextView
-        android:id="@+id/btc1W"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="BTC"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/eth1W" />
-
-    <TextView
-        android:id="@+id/btc1WPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="4dp"
-        android:text="-1,03%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/btc1W"
-        app:layout_constraintStart_toEndOf="@+id/btc1W"
-        app:layout_constraintTop_toTopOf="@+id/btc1W" />
-
-    <!--   Right   -->
-
-    <TextView
-        android:id="@+id/changes1M"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:text="Changes (1M)"
-        android:textColor="@color/grey"
-        android:textSize="14sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@id/middleGuideline"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <View
-        android:id="@+id/changes1MBorder"
-        android:layout_width="0dp"
-        android:layout_height="0.5dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="4dp"
-        android:background="@color/steel_20"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/middleGuideline"
-        app:layout_constraintTop_toBottomOf="@id/changes1M" />
-
-    <TextView
-        android:id="@+id/usd1M"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="8dp"
-        android:text="USD"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="@id/middleGuideline"
-        app:layout_constraintTop_toTopOf="@id/changes1MBorder" />
-
-    <TextView
-        android:id="@+id/usd1MPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="4dp"
-        android:text="+56,62%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/usd1M"
-        app:layout_constraintStart_toEndOf="@+id/usd1M"
-        app:layout_constraintTop_toTopOf="@+id/usd1M" />
-
-    <TextView
-        android:id="@+id/eth1M"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="8dp"
-        android:text="ETH"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="@id/middleGuideline"
-        app:layout_constraintTop_toBottomOf="@id/usd1M" />
-
-    <TextView
-        android:id="@+id/eth1MPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="4dp"
-        android:text="+56,62%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/eth1M"
-        app:layout_constraintStart_toEndOf="@+id/eth1M"
-        app:layout_constraintTop_toTopOf="@+id/eth1M" />
-
-    <TextView
-        android:id="@+id/btc1M"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="8dp"
-        android:text="BTC"
-        android:textColor="@color/bran"
-        android:textSize="14sp"
-        app:layout_constraintStart_toStartOf="@id/middleGuideline"
-        app:layout_constraintTop_toBottomOf="@id/eth1M" />
-
-    <TextView
-        android:id="@+id/btc1MPercent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="4dp"
-        android:text="+56,62%"
-        android:textColor="@color/remus"
-        android:textSize="12sp"
-        app:layout_constraintBottom_toBottomOf="@+id/btc1M"
-        app:layout_constraintStart_toEndOf="@+id/btc1M"
-        app:layout_constraintTop_toTopOf="@+id/btc1M" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
