+++ components/views/src/main/res/drawable-hdpi/rounded_steel10_background_bottom.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners
+        android:bottomLeftRadius="16dp"
+        android:bottomRightRadius="16dp" />
+    <solid android:color="@color/steel_10" />
+</shape>
